requirements.txt
streamlit
configparser
mysql-connector-python
pandas
plotly
matplotlib